fat-binary: Fix issue where fat binaries didn't pad all blocks.
For blocks before the last the EoF record bypassed the padding
of the last block, which affected alignment of fat binaries when
they were concatenated.

To solve this we simply remove the EoF record of the hex files,
except for the last one. The current algorith will pad all blocks
and and the last hex file will still have a shorter block without
padding and the End Of File record.